feat: Изменить тип данных chat_id на BigInteger для улучшения обработки идентификаторов

+++ app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship
 from app.core.database import Base
 
@@ -11,7 +11,7 @@ class User(Base):
     last_name = Column(String, nullable=False)
     role = Column(String, nullable=False)
     store_id = Column(Integer, ForeignKey("stores.id"), nullable=True)
-    chat_id = Column(Integer, unique=True, nullable=True)
+    chat_id = Column(BigInteger, unique=True, nullable=True)
 
     store = relationship("Store", back_populates="managers")
     revenues = relationship("Revenue", back_populates="manager")
